+++ docs/versionhistory.rst
@@ -4,6 +4,12 @@ Version history
 This library adheres to
 `Semantic Versioning 2.0 <https://semver.org/#semantic-versioning-200>`_.
 
+**UNRELEASED**
+
+- Fixed ``IndexError`` when using ``@typechecked`` on more than one function with the
+  same name under certain circumstances
+  (`#527 <https://github.com/agronholm/typeguard/issues/527>`_)
+
 **4.4.3** (2025-06-05)
 
 - Fixed ``@typechecked`` unable to find the target function or method if it or the